Fixed bug where game would not archive losses

+++ runapp.py
@@ -8,7 +8,7 @@ from abstractlevel.models import Game, PredArray
 # CREATE GAME
 #########################
 
-# Parameters
+# Settings
 numgames = 1000
 maxturns = 10**6
 nmonkeys = 1000
@@ -16,7 +16,8 @@ nsignals = 7
 nstates = 3
 minmonkeys = 30
 immortal = True
-archive_cycle = 1000
+archive_cycle = 10**4
+archive_loss = True
 
 predarray = PredArray([
     #grass  #tree   #bush
@@ -34,7 +35,8 @@ game = Game(
     mut_rate=0.05,
     min_monkeys=minmonkeys,
     immortal=immortal,
-    archive_cycle=archive_cycle)
+    archive_cycle=archive_cycle,
+    archive_loss=archive_loss)
 
 # CREATE ARCHIVE
 #########################
@@ -46,17 +48,26 @@ archive = pd.DataFrame(
         'Bottleneck Turn',
         'Worst Overall Turn Multiplier',
         'Best Overall Turn Multiplier',
+        'Optimal Response Chance',
+        'Losses',
     ]
 )
 
 if os.path.exists('archive.csv') and os.path.isfile('archive.csv'):
-    os.remove('archive.csv')
-
-archive.to_csv(
-    'archive.csv',
-    header=True,
-    index=False,
-    encoding='utf-8')
+    csv = pd.read_csv('archive.csv')
+    if (len(csv.columns) != len(archive.columns)) or (csv.columns != archive.columns).any():
+        os.remove('archive.csv')
+        archive.to_csv(
+            'archive.csv',
+            header=True,
+            index=False,
+            encoding='utf-8')
+else:
+    archive.to_csv(
+        'archive.csv',
+        header=True,
+        index=False,
+        encoding='utf-8')
 
 # RUN GAMES
 #########################
@@ -71,19 +82,19 @@ for i in range(numgames):
     bestgame = copy.deepcopy(game) if not bestgame else bestgame
     bestgame.numgame = i+1
     if game.monkeyswon:
-        print('MADE IT WITH %d MONKEYS! (bottleneck: %d monkeys in turn %d, bestmultiplier: %.4f, worstmultiplier: %.4f)' % (
+        print('MADE IT WITH %d MONKEYS!\n(bottleneck: %d monkeys in turn %d, bestmultiplier: %.4f, worstmultiplier: %.4f)' % (
             game.monkeyarray.nummonkeys,
             game.bottleneck,
             game.bottleneckturn,
             game.bestoverallturnmultiplier,
             game.worstoverallturnmultiplier))
     elif game.turns > bestgame.turns:
-        print('RECORD HIGH OF %d TURNS! (bestmultiplier: %.4f, worstmultiplier: %.4f)' % (
+        print('RECORD HIGH OF %d TURNS!\n(bestmultiplier: %.4f, worstmultiplier: %.4f)' % (
             game.turns,
             game.bestoverallturnmultiplier,
             game.worstoverallturnmultiplier))
     else:
-        print('%d TURNS. (bestmultiplier: %.4f, worstmultiplier: %.4f)' %(
+        print('%d TURNS.\n(bestmultiplier: %.4f, worstmultiplier: %.4f)' %(
             game.turns,
             game.bestoverallturnmultiplier,
             game.worstoverallturnmultiplier))
@@ -96,6 +107,8 @@ for i in range(numgames):
         'Bottleneck Turn': [game.bottleneckturn],
         'Worst Overall Turn Multiplier': [game.worstoverallturnmultiplier],
         'Best Overall Turn Multiplier': [game.bestoverallturnmultiplier],
+        'Optimal Response Chance': [str(game.optimalchance)],
+        'Losses': [game.losses],
     })
 
     archive.to_csv(
